New pipeline for skystone detection
This is the new DogeCV pipeline and opmode template for skystone detection credit to Easy Open CV. This works much faster than vuforia and is much more accurate. Changes made to this from original include  passing a webcam instead of phone cam, new imports, and changes to the instantiation of the skystone detector before calling useDefaults( ) function.